+++ src/loaders.py
@@ -25,9 +25,7 @@ def save_list_of_dicts_to_dataframe(dict_list: list[dict], save_options: dict, d
     filename = f"data.{format}"
     save_path = os.path.join(folder, dt, filename)
     if format == "csv":
-        df.to_csv(
-            save_path,
-        )
+        df.to_csv(save_path, index=False)
     else:
         # by default save parquet unless otherwise
         df.to_parquet(save_path)
